pyproject.toml
[project]
name = "eftdiscord"
version = "0.1.0"
description = ""
authors = [
    {name = "LedxDelivery"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "pypresence @ http://github.com/qwertyquerty/pypresence/archive/master.zip",
    "psutil (>=7.1.3,<8.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyinstaller (>=6.17.0,<7.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pycryptodome-test-vectors (>=1.0.22,<2.0.0)",
    "pydoll-python (>=2.12.4,<3.0.0)",
    "pyprof2calltree (>=1.4.5,<2.0.0)",
    "snakeviz (>=2.2.2,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
